requirements.txt
absl-py==0.10.0
aiohttp==3.7.2
aiohttp-cors==0.7.0
aioredis==1.3.1
appdirs==1.4.4
appnope==0.1.0
argo-models==2.2.1a0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrdict==2.0.1
attrs==20.2.0
backcall==0.1.0
bayesian-optimization==1.2.0
beautifulsoup4==4.9.3
black==20.8b1
bleach==3.2.1
blessings==1.7
boto==2.49.0
cachetools==4.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
colorful==0.5.4
contextlib2==0.6.0.post1
cryptography==3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
diffprivlib==0.3.0
dm-tree==0.1.5
docutils==0.16
entrypoints==0.3
filelock==3.0.12
fsspec==0.7.4
future==0.18.2
gast==0.3.3
gcsfs==0.6.2
google==3.0.0
google-api-core==1.16.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-pasta==0.1.8
google-resumable-media==0.5.1
googleapis-common-protos==1.51.0
gpustat==0.6.0
grpcio==1.33.2
h5py==2.10.0
hiredis==1.1.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
ipyhc==0.0.0a5
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.4.0
kfp==0.4.0
kfp-server-api==0.3.0
kiwisolver==1.2.0
kubernetes==11.0.0
loguru==0.5.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
more-itertools==8.5.0
mpmath==1.1.0
msgpack==1.0.0
multidict==5.0.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.0.3
numpy==1.19.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opacus==0.9.1
opencensus==0.7.11
opencensus-context==0.1.2
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipdeptree==1.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.7
protobuf==3.11.3
psutil==5.7.3
ptyprocess==0.6.0
py==1.9.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.15.7
pytest==6.0.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.6.0
ray==1.0.0
readme-renderer==26.0
redis==3.4.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.22.2.post1
scipy==1.5.2
SecretStorage==3.1.2
Send2Trash==1.5.0
sentencepiece==0.1.91
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
statsmodels==0.11.1
strip-hints==0.1.9
sympy==1.5.1
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow-estimator==2.3.0
tensorflow-privacy==0.5.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tokenizers==0.9.2
toml==0.10.1
torch==1.6.0
torchcsprng==0.1.2
torchtext==0.7.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
transformers==3.4.0
twine==3.2.0
typer==0.3.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.2
zipp==3.1.0
